Initial commit: a working clock counting only seconds
My work so far. Each LED is connected to a pin on an Arduino. First we
get the seconds part of the current time, then convert it to binary and
use that to send the HIGH or LOW signals to the pins to show the time
(just the seconds) on the LEDs. This is repeated as fast as possible.

Most of the code was originally written in C on a train and then
modified to run on the Arduino as soon as I had one to test it on.