Fix dependency in header file
It's been a while since I wrote C. Apparently I need to include the
binclock C source file in the binclock header file so that the functions
are callable when including the header file from other files. This also
means I can put my struct in the C file where it belongs and the code
compiles!